Add files via upload
A cooperative of farmers requires a system that advises which cereal can be planted in a given lot. The lots are classified as special when they contain certain minerals of primary interest to the cooperative and common when they contain only secondary minerals in the composition of the land.
In order to plant a cereal in a field, it must contain certain minerals. Cereals are classified into coarse crop grains (such as sunflower, corn,...), fine crop grains (such as wheat, oats,...) and pastures (such as alfalfa, subterranean clover,...). A lot satisfies the requirements of a cereal if it contains all of the minerals required by that cereal. In addition, in the case of pastures, another pasture cannot have been previously planted.